pyproject.toml
[project]
name = "langraph-agent-initial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=0.3.5",
    "langgraph>=0.2.23",
    "langgraph-checkpoint-sqlite>=1.0.3",
    "langgraph-sdk>=0.1.30",
    "langsmith>=0.1.125",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "tavily-python>=0.5.0",
    "wikipedia>=1.4.0",
    "langchain-groq>=0.2.0",
    "google-generativeai==0.7.1",
    "langchain-google-genai>=2.0.0",
]
